+++ builds/4chan-XT.user.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name         4chan XT
-// @version      2.16.0
+// @version      2.17.0
 // @minGMVer     1.14
 // @minFFVer     74
 // @namespace    4chan-XT
@@ -169,8 +169,8 @@
   'use strict';
 
   var version = {
-    "version": "2.16.0",
-    "date": "2024-10-26T11:08:00Z"
+    "version": "2.17.0",
+    "date": "2024-11-10T18:28:00Z"
   };
 
   var meta = {
@@ -4928,7 +4928,7 @@ $site$infoRoot a.hide-reply-button {
 .stub input {
   display: inline-block;
 }
-.stub-icon,
+.stub .stub-icon,
 .stub-subject {
   margin-right: 1ch;
 }
@@ -10661,35 +10661,45 @@ svg.icon {
 
     cb: {
       openAll() {
-        if ($.hasClass(this, 'disabled')) { return; }
+        if ($.hasClass(this, 'disabled')) return;
         for (var a of $$('a.watcher-link', ThreadWatcher.list)) {
           $.open(a.href);
         }
-        return $.event('CloseMenu');
+        $.event('CloseMenu');
       },
       openUnread() {
-        if ($.hasClass(this, 'disabled')) { return; }
+        if ($.hasClass(this, 'disabled')) return;
         for (var a of $$('.replies-unread > a.watcher-link', ThreadWatcher.list)) {
           $.open(a.href);
         }
-        return $.event('CloseMenu');
+        $.event('CloseMenu');
       },
       openDeads() {
-        if ($.hasClass(this, 'disabled')) { return; }
-        for (var a of $$('.dead-thread > a.watcher-link', ThreadWatcher.list)) {
+        if ($.hasClass(this, 'disabled')) return;
+        for (var a of $$('.dead-thread.replies-unread > a.watcher-link', ThreadWatcher.list)) {
           $.open(a.href);
         }
-        return $.event('CloseMenu');
+        $.event('CloseMenu');
       },
       pruneDeads() {
-        if ($.hasClass(this, 'disabled')) { return; }
+        if ($.hasClass(this, 'disabled')) return;
         for (var {siteID, boardID, threadID, data} of ThreadWatcher.getAll()) {
           if (data.isDead) {
             ThreadWatcher.db.delete({siteID, boardID, threadID});
           }
         }
         ThreadWatcher.refresh();
-        return $.event('CloseMenu');
+        $.event('CloseMenu');
+      },
+      pruneReadDeads() {
+        if ($.hasClass(this, 'disabled')) return;
+        for (var { siteID, boardID, threadID, data } of ThreadWatcher.getAll()) {
+          if (data.isDead && !data.unread) {
+            ThreadWatcher.db.delete({ siteID, boardID, threadID });
+          }
+        }
+        ThreadWatcher.refresh();
+        $.event('CloseMenu');
       },
       dismiss() {
         for (var {siteID, boardID, threadID, data} of ThreadWatcher.getAll()) {
@@ -10697,26 +10707,26 @@ svg.icon {
             ThreadWatcher.update(siteID, boardID, threadID, {dismiss: data.quotingYou || 0});
           }
         }
-        return $.event('CloseMenu');
+        $.event('CloseMenu');
       },
       toggle() {
         const {thread} = Get.postFromNode(this);
-        return ThreadWatcher.toggle(thread);
+        ThreadWatcher.toggle(thread);
       },
       rm() {
         const {siteID} = this.parentNode.dataset;
         const [boardID, threadID] = this.parentNode.dataset.fullID.split('.');
-        return ThreadWatcher.rm(siteID, boardID, +threadID);
+        ThreadWatcher.rm(siteID, boardID, +threadID);
       },
       post(e) {
         const {boardID, threadID, postID} = e.detail;
         const cb = PostRedirect.delay();
         if (postID === threadID) {
           if (Conf['Auto Watch']) {
-            return ThreadWatcher.addRaw(boardID, threadID, {}, cb);
+            ThreadWatcher.addRaw(boardID, threadID, {}, cb);
           }
         } else if (Conf['Auto Watch Reply']) {
-          return ThreadWatcher.add((g.threads.get(boardID + '.' + threadID) || new Thread(threadID, g.boards[boardID] || new Board(boardID))), cb);
+          ThreadWatcher.add((g.threads.get(boardID + '.' + threadID) || new Thread(threadID, g.boards[boardID] || new Board(boardID))), cb);
         }
       },
       onIndexUpdate(e) {
@@ -11051,8 +11061,7 @@ svg.icon {
       const x = $.el('a', {
         textContent: '✕',
         href: 'javascript:;'
-      }
-      );
+      });
       $.on(x, 'click', ThreadWatcher.cb.rm);
 
       let {excerpt, isArchived} = data;
@@ -11063,15 +11072,13 @@ svg.icon {
         href: g.sites[siteID]?.urls.thread({siteID, boardID, threadID}, isArchived) || '',
         title: excerpt,
         className: 'watcher-link'
-      }
-      );
+      });
 
       if (Conf['Show Page'] && (data.page != null)) {
         page = $.el('span', {
           textContent: `[${data.page}]`,
           className: 'watcher-page'
-        }
-        );
+        });
         $.add(link, page);
       }
 
@@ -11079,16 +11086,14 @@ svg.icon {
         const count = $.el('span', {
           textContent: `(${data.unread})`,
           className: 'watcher-unread'
-        }
-        );
+        });
         $.add(link, count);
       }
 
       const title = $.el('span', {
         textContent: excerpt,
         className: 'watcher-title'
-      }
-      );
+      });
       $.add(link, title);
 
       const div = $.el('div');
@@ -11323,24 +11328,30 @@ svg.icon {
           }
         });
 
-        // `Open dead threads` entry
+        const toggleDisabledDead = function () {
+          this.el.classList.toggle('disabled', !$('.dead-thread', ThreadWatcher.list));
+          return true;
+        };
+
+        // `Open unread dead threads` entry
         entries.push({
-          text: 'Open dead threads',
+          text: 'Open unread dead threads',
           cb: ThreadWatcher.cb.openDeads,
-          open() {
-            this.el.classList.toggle('disabled', !$('.dead-thread', ThreadWatcher.list));
-            return true;
-          }
+          open: toggleDisabledDead,
         });
 
-        // `Prune dead threads` entry
+        // `Prune all dead threads` entry
         entries.push({
-          text: 'Prune dead threads',
+          text: 'Prune all dead threads',
           cb: ThreadWatcher.cb.pruneDeads,
-          open() {
-            this.el.classList.toggle('disabled', !$('.dead-thread', ThreadWatcher.list));
-            return true;
-          }
+          open: toggleDisabledDead,
+        });
+
+        // `Prune read dead threads` entry
+        entries.push({
+          text: 'Prune read dead threads',
+          cb: ThreadWatcher.cb.pruneReadDeads,
+          open: toggleDisabledDead,
         });
 
         // `Dismiss posts quoting you` entry
@@ -11359,8 +11370,7 @@ svg.icon {
             el: $.el('a', {
               textContent: text,
               href: 'javascript:;'
-            }
-            )
+            })
           };
           if (title) { entry.el.title = title; }
           $.on(entry.el, 'click', cb);
@@ -15595,7 +15605,7 @@ aero|asia|biz|cat|com|coop|dance|info|int|jobs|mobi|moe|museum|name|net|org|post
       },
       {
         key: 'Twitter',
-        regExp: /^\w+:\/\/(?:www\.|mobile\.)?(?:twitter|x)\.com\/(\w+\/status\/\d+)/,
+        regExp: /^\w+:\/\/(?:www\.|mobile\.)?(?:(?:(?:fx|vx)?twitter|(?:fixup|fixv)?x|twittpr|xcancel)\.com|nitter\.\w+.\w+)\/(\w+\/status\/\d+)/,
         style: 'border: none; width: 550px; height: 250px; overflow: hidden; resize: both;',
         el(a) {
           if (Conf.XEmbedder === 'tf') {